Relocate override config functions
