+++ tests/integration/test_authd/force_options/test_authd_force_options.py
@@ -3,7 +3,7 @@ import time
 import pytest
 import yaml
 from wazuh_testing.tools.monitoring import make_callback, AUTHD_DETECTOR_PREFIX
-from wazuh_testing.tools.configuration import load_wazuh_configurations, set_section_wazuh_conf
+from wazuh_testing.tools.configuration import load_wazuh_configurations
 from wazuh_testing.tools.file import read_yaml
 from wazuh_testing.authd import create_authd_request, validate_authd_response, AUTHD_KEY_REQUEST_TIMEOUT, \
                                 insert_pre_existent_agents
@@ -44,19 +44,7 @@ def validate_authd_logs(logs):
                           error_message=f"Expected error log does not occured: '{log}'")
 
 
-def create_force_config_block(param):
-    """
-    Creates a temporal config file.
-    """
-    temp = os.path.join(data_path, 'temp.yaml')
 
-    with open(configurations_path, 'r') as conf_file:
-        temp_conf_file = yaml.safe_load(conf_file)
-        for elem in param:
-            temp_conf_file[0]['sections'][0]['elements'].append(elem)
-    with open(temp, 'w') as temp_file:
-        yaml.safe_dump(temp_conf_file, temp_file)
-    return temp
 
 
 # Fixtures
@@ -74,24 +62,6 @@ def get_current_test_case(request):
     Get current test case from the module
     """
     return request.param
-
-
-@pytest.fixture(scope='function')
-def format_configuration(get_current_test_case, request):
-    """
-    Get configuration block from current test case
-    """
-    test_name = request.node.originalname
-    configuration = get_current_test_case.get('configuration', {})
-
-    # Configuration for testing
-    temp = create_force_config_block(configuration)
-    conf = load_wazuh_configurations(temp, test_name)
-    os.remove(temp)
-
-    test_config = set_section_wazuh_conf(conf[0]['sections'])
-
-    return test_config
 
 
 def test_authd_force_options(get_current_test_case, configure_local_internal_options_module, override_authd_force_conf,
